The moment I realized I need Redux!
Check out MenuSelectItem.js and run the app. Notice that the grey bar
moves around, but there’s an error about render needing to be ‘pure’.

There are at least two problems with this implementation. The first is
that I’m modifying state in render, which is not allowed in react. The
second is that the rest of the application (routing, history, other
components) will have no idea what MenuListItem the user has chosen.

I have another problem too. I need the KeyboardEvents component to act
as an event emitter, but I haven’t figured out a great way to listen to
the events I emit.
I need something that can keep track of a global state no matter what
the views are, something that any component can subscribe to for state
changes, and something that any component can use to change the
applications state.

Enter redux! It does these three things.